requirements.txt
PyMuPDF
opencv-python
numpy
pytesseract
Pillow
matplotlib
transformers
torch